fastapi/coordinator/API/requirements.txt
fastapi==0.78.0
uvicorn[standard]==0.18.2
SQLAlchemy==1.4.39
psycopg2-binary==2.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
flake8==4.0.1
requests==2.28.1
paramiko==2.11.0
typing==3.7.4.3
aio-pika==8.1.1
python-powerdns==2.1.0
cryptography==37.0.4
fastapi_keycloak==1.0.7